feat(web): create new user api server functions
