+++ apps/web/src/app/(pages)/(auth)/types.ts
@@ -2,4 +2,8 @@ export interface User {
   id: string;
   email: string;
   picture: string;
+
+  username?: string;
+  firstName?: string;
+  lastName?: string;
 }
