+++ apps/web/src/app/_common/actions/auth.action.ts
@@ -3,10 +3,9 @@
 import * as Sentry from '@sentry/nextjs';
 import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
-import { type User } from '~/app/(pages)/(auth)/types';
 import { type TokenResponse } from '../interfaces/token.interface';
-import { getAuthCookie } from '../utils/get-auth-cookie';
 import { http } from '../utils/http';
+import { getMe } from './user.action';
 
 const ACCESS_TOKEN_COOKIE_NAME = 'access_token';
 const REFRESH_TOKEN_COOKIE_NAME = 'refresh_token';
@@ -41,22 +40,6 @@ export const refreshNewAccessToken = async () => {
       await signOut();
     }
 
-    return null;
-  }
-};
-
-export const getMe = async (): Promise<User | null> => {
-  try {
-    const user: User = await http
-      .get('users/me', {
-        headers: {
-          Cookie: await getAuthCookie(),
-        },
-      })
-      .json();
-    return user;
-  } catch (error) {
-    Sentry.captureException(error);
     return null;
   }
 };
