+++ apps/web/src/app/_common/components/top-nav-bar.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
-import { getMe } from '../actions/auth.action';
 import { getCategories } from '../actions/category.action';
+import { getMe } from '../actions/user.action';
 import { AnimatedTab } from './animated-tab';
 import { Logo } from './logo';
 import { UserAvatar } from './user-avatar';
